Release v3.7.4 (20240308)

+++ test/api_atcab/atca_tests_buffer.c
@@ -39,13 +39,13 @@ void init_ref_array(uint8_t * buf, size_t len)
 
     for (i = 0U; i < len; i++)
     {
-        buf[i] = i;
+        buf[i] = (uint8_t)i;
     }
 }
 
 static unsigned int check_ref_array(uint8_t * buf, size_t len, size_t offset)
 {
-    unsigned int i;
+    size_t  i;
 
     for (i = offset; i < len; i++)
     {
@@ -54,7 +54,7 @@ static unsigned int check_ref_array(uint8_t * buf, size_t len, size_t offset)
             break;
         }
     }
-    return i - offset;
+    return (unsigned int)(i - offset);
 }
 
 static unsigned int check_array(uint8_t* buf, size_t len, uint8_t val)
